Split tv and iPhone layouts
